Added first publicly viewable version of the overhaul.
Assets nowhere near finished, obviously.
If anyone who finds this finds any bugs/issues, please post them in the issues section, include what build you are using by date.

Do not reupload anywhere. All assets on this git are a work in progress, and not intended as a final product.